requirements.txt
Django>=3.2
djangorestframework>=3.12
psycopg2-binary>=2.8